use more detailed names for the organisms
From Suzie:

As of now, this program simply prints out the full name from the
GenBank description The first word is often the species, sometimes the
second word is an additional identifier. We want to compare the
species names, but also organize them by the subspecies (second name)
whenever possible for identifying non-monophyly.

The reason for this script is the FASTA names are sometimes not
descriptive enough, and some individuals appear as the same species
when in fact we want to see the non-monophyly occurring on a sub-
species level, or sometimes the wrong name is taken and we see
individuals of the  same species identified as different.